Removed unnecessary transfer specific structs
