+++ src/models/backend/mod.rs
@@ -1,6 +1,5 @@
 pub mod about;
 pub mod balances;
-pub mod multisig_transaction_request;
 pub mod transactions;
 pub mod transfers;
 pub mod webhooks;
