--- src/models/backend/multisig_transaction_request.rs
@@ -1,20 +1,0 @@
-use crate::models::commons::Operation;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct MultisigTransactionRequest {
-    pub to: String,
-    pub value: String,
-    pub data: String,
-    pub nonce: String,
-    pub operation: Operation,
-    pub safe_tx_gas: String,
-    pub base_gas: String,
-    pub gas_price: String,
-    pub gas_token: String,
-    pub refund_receiver: String,
-    pub contract_transaction_hash: String,
-    pub sender: String,
-    pub signature: String,
-}
