+++ src/routes/mod.rs
@@ -25,7 +25,7 @@ pub fn active_routes() -> Vec<Route> {
         transactions::history_transactions,
         transactions::queued_transactions,
         transactions::submit_confirmation,
-        transactions::send_eth,
+        transactions::propose_transaction,
         hooks::update,
         health::health
     ]
