+++ src/routes/transactions.rs
@@ -1,6 +1,7 @@
 use crate::config::request_cache_duration;
-use crate::models::service::transactions::requests::{ConfirmationRequest, SendFundsRequest};
-use crate::services::tx_confirmation;
+use crate::models::service::transactions::requests::{
+    ConfirmationRequest, MultisigTransactionRequest,
+};
 use crate::services::{
     transactions_details, transactions_history, transactions_list, transactions_proposal,
     transactions_queued,
@@ -42,7 +43,7 @@ pub fn submit_confirmation(
     safe_tx_hash: String,
     tx_confirmation_request: Json<ConfirmationRequest>,
 ) -> ApiResult<content::Json<String>> {
-    tx_confirmation::submit_confirmation(
+    transactions_proposal::submit_confirmation(
         &context,
         &safe_tx_hash,
         &tx_confirmation_request.signed_safe_tx_hash,
@@ -97,13 +98,17 @@ pub fn queued_transactions(
 }
 
 #[post(
-    "/v1/transactions/<safe_address>/transfers",
-    data = "<send_funds_request>"
+    "/v1/transactions/<safe_address>/propose",
+    data = "<multisig_transaction_request>"
 )]
-pub fn send_eth(
+pub fn propose_transaction(
     context: Context,
     safe_address: String,
-    send_funds_request: Json<SendFundsRequest>,
+    multisig_transaction_request: Json<MultisigTransactionRequest>,
 ) -> ApiResult<()> {
-    transactions_proposal::send_funds(&context, &safe_address, &send_funds_request)
+    transactions_proposal::propose_transaction(
+        &context,
+        &safe_address,
+        &multisig_transaction_request,
+    )
 }
