+++ src/services/mod.rs
@@ -9,7 +9,6 @@ pub mod transactions_history;
 pub mod transactions_list;
 pub mod transactions_proposal;
 pub mod transactions_queued;
-pub mod tx_confirmation;
 
 #[cfg(test)]
 mod tests;
